[Профиль] Поправил форматирование в файлах эпика.
