+++ FakeNFT/Epics/Profile/Helpers/ProfileNetworkClient.swift
@@ -2,12 +2,15 @@ import Foundation
 
 final class ProfileNetworkClient: ProfileNetworkClientProtocol {
     // MARK: - Public properties
+    
     weak var presenter: ProfilePresenterNetworkProtocol?
     
     // MARK: - Private properties
+    
     private let networkClient = DefaultNetworkClient()
     
     // MARK: - Public properties
+    
     func getDecodedProfile(){
         networkClient.send(request: ProfileRequest(), type: ProfileResponseModel.self){ result in
             switch result {
@@ -30,26 +33,5 @@ final class ProfileNetworkClient: ProfileNetworkClientProtocol {
                 completion(.failure(error))
             }
         }
-    }
-}
-
-
-// MARK: - Requests
-struct  ProfileRequest: NetworkRequest {
-    var endpoint: URL? {
-        return URL(string: ProfileConstants.endpoint + ProfileConstants.apiV1Profile1Get)
-    }
-}
-
-struct ProfileUpdateRequest: NetworkRequest {
-    var endpoint: URL? {
-        return URL(string: ProfileConstants.endpoint + ProfileConstants.apiV1Profile1Get)
-    }
-    var httpMethod: HttpMethod {
-        return .put
-    }
-    var dto: Encodable?
-    init(updatedData: ProfileResponseModel) {
-        self.dto = updatedData
     }
 }
