+++ FakeNFT/Epics/Profile/Presentation/FavoriteNFTs/CollectionParameters.swift
@@ -2,6 +2,7 @@ import UIKit
 
 struct FeaturedCollectionParameters {
     // MARK: - Public proprties
+    
     let cellCount: Int
     let leftInset: CGFloat
     let rightInset: CGFloat
@@ -9,7 +10,13 @@ struct FeaturedCollectionParameters {
     let paddingWidth: CGFloat
     
     // MARK: - Life cycle
-    init(cellCount: Int, leftInset: CGFloat, rightInset: CGFloat, cellSpacing: CGFloat) {
+    
+    init(
+        cellCount: Int,
+        leftInset: CGFloat,
+        rightInset: CGFloat,
+        cellSpacing: CGFloat
+    ) {
         self.cellCount = cellCount
         self.leftInset = leftInset
         self.rightInset = rightInset
