+++ FakeNFT/Epics/Profile/Presentation/ProfileSummary/ProfileTableViewCell.swift
@@ -2,6 +2,7 @@ import UIKit
 
 final class ProfileTableViewCell: UITableViewCell, ReuseIdentifying {
     // MARK: - Private properties
+    
     private let titleLabel = UILabel()
     private let chevronView = UIImageView()
     private let stackView: UIStackView = {
@@ -14,6 +15,7 @@ final class ProfileTableViewCell: UITableViewCell, ReuseIdentifying {
     }()
     
     // MARK: - Life cicle
+    
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         self.selectionStyle = .none
@@ -32,11 +34,13 @@ final class ProfileTableViewCell: UITableViewCell, ReuseIdentifying {
     }
     
     // MARK: - Public methods
+    
     func configure(title: String) {
         configureTitleLabel(with: title)
     }
     
     // MARK: - Private methods
+    
     private func layoutConfigure() {
         contentView.addSubview(stackView)
