+++ FakeNFT/Epics/Profile/Protocols/ProfileEditProtocols.swift
@@ -1,6 +1,7 @@
 import Foundation
 
 // MARK: - ProfileEditPresenterProtocol
+
 protocol ProfileEditPresenterProtocol: AnyObject {
     func change(parameter: EditableParameter, with text: String)
     func isProfileChanged() -> Bool
